requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
coverage==5.5
Django==2.2.19
django-ckeditor==6.0.0
django-js-asset==1.2.2
factory-boy==3.2.0
Faker==8.9.0
inflection==0.5.1
iniconfig==1.1.1
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
